Add files via upload
Read through the transcript of our call with the Hong Kong Director as well as my notes from the call that provide an overview of the targets. You can find these attached below.

Following this, consider all five targets discussed and email Anna with a concise description (1-2 sentences) of each company and why they are (or are not) appropriate to share with Carlos as potential M&A targets. Please use the email template provided below.

Note: when deciding if a target is appropriate to share with Carlos, you can consider the following elements:

Does this make sense from a strategic and financial perspective - consider which acquisitions may be a good fit for your client
Given the shareholder structures of the targets, consider the simplicity/feasibility of each option